change default dtype in NumpyDataset to uint8, allow setting dtype by user

+++ tests/conftest.py
@@ -60,7 +60,9 @@ def test_numpy_dataset():
     )
     sample_ids = pd.Series(["sample1", "sample2", "sample3"])
     populations = pd.Series(["population1", "population2", "population3"])
-    dataset = NumpyDataset(test_data, sample_ids, populations, missing_value=0)
+    dataset = NumpyDataset(
+        test_data, sample_ids, populations, missing_value=0, dtype=np.uint8
+    )
     return dataset
